+++ R/format.R
@@ -57,16 +57,11 @@ print.interval <- function(x, digits = NULL, ...) {
 
 #' @export
 format.interval <- function(x, digits = NULL, ...) {
-  if (is_empty(x)) {
-    return("!")
-  }
+  if (is_empty(x)) return("!")
   not_zero <- !purrr::map_lgl(x, function(x) x == 0)
-  # if output is empty, it means that duplicated time entries
-  # if output is NA, it means that only one time entry
+  # if output contains all the zeros
+  if (sum(not_zero) == 0) return("?")
   output <- x[not_zero]
   vec_x <- rlang::flatten_dbl(output)
-  if (is_empty(output) || is_empty(vec_x)) {
-    return("?")
-  }
   paste0(vec_x, toupper(names(output)), collapse = " ")
 }