+++ R/interval.R
@@ -21,13 +21,7 @@
 #' pull_interval(x)
 pull_interval <- function(x) {
   if (has_length(x, 1)) {
-    return(structure(
-      list(
-        year = NA, quarter = NA, month = NA, week = NA,
-        hour = NA, minute = NA, second = NA, unit = NA
-      ),
-      class = "interval"
-    ))
+    return(init_interval())
   }
   UseMethod("pull_interval")
 }
@@ -38,13 +32,10 @@ pull_interval.POSIXt <- function(x) {
   dttm <- as.numeric(x)
   nhms <- gcd_interval(dttm) # num of seconds
   period <- split_period(nhms)
-  structure(
-    list(
-      hour = period$hour, 
-      minute = period$minute, 
-      second = round(period$second, digits = 3)
-    ),
-    class = "interval"
+  init_interval(
+    hour = period$hour, 
+    minute = period$minute, 
+    second = round(period$second, digits = 3)
   )
 }
 
@@ -53,7 +44,7 @@ pull_interval.nanotime <- function(x) {
   nano <- as.numeric(x)
   int <- gcd_interval(nano) # num of nanoseconds
   sec <- int * 1e-9
-  structure(list(second = sec), class = "interval")
+  init_interval(second = sec)
 }
 
 #' @export
@@ -66,21 +57,21 @@ pull_interval.hms <- pull_interval.difftime # for hms package
 pull_interval.Date <- function(x) {
   dttm <- as.numeric(x)
   ndays <- gcd_interval(dttm) # num of seconds
-  structure(list(day = ndays), class = "interval")
+  init_interval(day = ndays)
 }
 
 #' @export
 pull_interval.yearweek <- function(x) {
   wk <- units_since(x)
   nweeks <- gcd_interval(wk)
-  structure(list(week = nweeks), class = "interval")
+  init_interval(week = nweeks)
 }
 
 #' @export
 pull_interval.yearmonth <- function(x) {
   mon <- units_since(x)
   nmonths <- gcd_interval(mon)
-  structure(list(month = nmonths), class = "interval")
+  init_interval(month = nmonths)
 }
 
 #' @export
@@ -92,7 +83,7 @@ pull_interval.yearmth <- function(x) {
 pull_interval.yearquarter <- function(x) {
   qtr <- units_since(x)
   nqtrs <- gcd_interval(qtr)
-  structure(list(quarter = nqtrs), class = "interval")
+  init_interval(quarter = nqtrs)
 }
 
 #' @export
@@ -104,9 +95,9 @@ pull_interval.yearqtr <- function(x) {
 pull_interval.numeric <- function(x) {
   nunits <- gcd_interval(x)
   if (min0(x) > 1599 && max0(x) < 2500) {
-    return(structure(list(year = nunits), class = "interval"))
+    return(init_interval(year = nunits))
   }
-  structure(list(unit = nunits), class = "interval")
+  init_interval(unit = nunits)
 }
 
 #' @export
@@ -117,6 +108,17 @@ pull_interval.numeric <- function(x) {
 #' @export
 `[.interval` <- function(x, i, j, drop = FALSE) {
   NextMethod()
+}
+
+init_interval <- function(
+  year = 0, quarter = 0, month = 0, week = 0, 
+  day = 0, hour = 0, minute = 0, second = 0, unit = 0
+) {
+  structure(list(
+    year = year, quarter = quarter, month = month, week = week,
+    day = day, hour = hour, minute = minute, second = second,
+    unit = unit
+  ), class = "interval")
 }
 
 #' @rdname pull-interval