+++ R/utils.R
@@ -74,10 +74,6 @@ dont_know <- function(x, FUN) {
 }
 
 unknown_interval <- function(x) {
-  not_zero <- !purrr::map_lgl(x, function(x) x == 0)
-  # if output is empty, it means that duplicated time entries
-  # if output is NA, it means that only one time entry
-  if (anyNA(not_zero)) {
-    abort("Cannot deal with data of unknown interval.")
-  }
+  no_zeros <- !purrr::map_lgl(x, function(x) x == 0)
+  if (sum(no_zeros) == 0) abort("Cannot deal with data of unknown interval.")
 }